src/components/image_analyzer/service1/requirements.txt
fastapi
uvicorn
pydantic
pillow
rembg
torch
transformers